Collapse multiple packages into core package
I'm not entirely convinced this is the best thing to do, but I found the
current separation too artificial, so we'll go with one major package to
implement the core solver functionality and one separate package for
integration tests for now. If the need arises to gain more control over
visibility of names to enable stricter encapsulation, we can revisit
this.

The module 'deflect' now has a package 'deflect', which seems a bit
silly, but it doesn't seem to be bad practice either (at least I
couldn't find anything on that).