Add type and interface for element Neumann BC
The opaque any type might look like a weird choice, so let me explain.
An element implementation and the Neumann BC that can be applied to it
are tightly coupled. There is no meaningful way to separate those two by
means of an interface, e.g., for element loads (an Apply/Assemble
function similar to the Element's one), it always ends up with more
complexity than justified. And again, this is to be expected, because
element formulation and element load are the same concept, since the
interpolation function determines how the element loads are attributed
to nodes.

Element formulations are hidden behind the Element interface from the
moment on where they are constructed. Since the Element interface is
generic to all elements (a truss, a spring, a frame, ...), it can't have
a function to add an element-specific load object, since we want to be
able to add new load types without having to tweak the stable Element
interface. A consistent approach would be to pass the loads upon
construction, when the concrete element type is known. However, this is
not practical because any workflow/UI will allow for interactive
addition and removal of element loads, so this should be modeled in the
interface, too.

Another option is to store element loads separately from the elements.
However, this again requires an abstraction type for element loads,
which effectively shouldn't be abstracted away from the element
implementations.

We choose a pragmatic compromise here. The most generic type is used to
represent an element load, and when the element assembles global tangent
and residual or computes stress interpolations, it is also responsible
for downcasting these loads and use the resulting object to do this. As
a benefit, we have one place where coordinate transformation, static
condensation of internal degrees of freedom (e.g., due to hinges) is
accounted for and there is no need for complicated orchestration of the
dependencies between these element internals and the element loading.