Implement map utility
The filename might not be ideal; as if we'd try to mirror C++ here. I
had thought about functional.go, but the implementation style here is
only slightly functional, since the transformation happens in place,
rather than returning a function that performs the transformation when
invoked. Depending on what will be added to this file, we might want to
revisit its name.