Disallow unknown 'expected' JSON subfields
This makes it easier to detect tests with typos or other accidental JSON
fields under the expected object. The problem was that any such mistake
would lead to a test that would silently not execute the faulty
expectation specification, leading to a potential false positive test
success.

Includes a mild refactoring of error descriptors etc.